pyproject.toml
[tool.poetry]
name = "openai-experiments"
version = "0.1.0"
description = "Reproducible experiments with OpenAI models"
authors = ["Adriano Valdes-Gomez <adriano@ciencias.unam.mx>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"